Extract CLI startup context builder
